fix: post/put issues, added delete function, fixed enums

+++ project2_personal_portfolio_tracker/src/main/java/com/skillstorm/Repositories/InvestmentAccountRepo.java
@@ -8,6 +8,6 @@ import com.skillstorm.Models.InvestmentAccount;
 
 public interface InvestmentAccountRepo extends JpaRepository<InvestmentAccount, Integer>{
 
-    Iterable<InvestmentAccount> findByUserId(Long userId);
+    Iterable<InvestmentAccount> findByUserId(int userId);
     
 }
